Fix all the bugs introduced by the cherry picking

+++ libs/gui/GraphicBufferAlloc.cpp
@@ -35,7 +35,7 @@ GraphicBufferAlloc::~GraphicBufferAlloc() {
 sp<GraphicBuffer> GraphicBufferAlloc::createGraphicBuffer(uint32_t width,
         uint32_t height, PixelFormat format, uint32_t usage, status_t* error) {
     sp<GraphicBuffer> graphicBuffer(
-            new GraphicBuffer(width, height, format, usage, mBufferSize));
+            new GraphicBuffer(width, height, format, usage, static_cast<uint32_t>(mBufferSize)));
     status_t err = graphicBuffer->initCheck();
     *error = err;
     if (err != 0 || graphicBuffer->handle == 0) {
